Flutter Driver: command extensions and extension feature cleanup (#67916)
