+++ packages/flutter_driver/lib/driver_extension.dart
@@ -24,5 +24,6 @@
 ///     }
 library flutter_driver_extension;
 
-export 'src/common/create_finder_factory.dart';
+export 'src/common/deserialization_factory.dart';
+export 'src/common/handler_factory.dart';
 export 'src/extension/extension.dart';
