+++ packages/flutter_driver/lib/src/common/find.dart
@@ -6,27 +6,9 @@ import 'dart:convert';
 
 import 'package:meta/meta.dart';
 
+import 'deserialization_factory.dart';
 import 'error.dart';
 import 'message.dart';
-
-/// A factory for deserializing [Finder]s.
-mixin DeserializeFinderFactory {
-  /// Deserializes the finder from JSON generated by [SerializableFinder.serialize].
-  SerializableFinder deserializeFinder(Map<String, String> json) {
-    final String? finderType = json['finderType'];
-    switch (finderType) {
-      case 'ByType': return ByType.deserialize(json);
-      case 'ByValueKey': return ByValueKey.deserialize(json);
-      case 'ByTooltipMessage': return ByTooltipMessage.deserialize(json);
-      case 'BySemanticsLabel': return BySemanticsLabel.deserialize(json);
-      case 'ByText': return ByText.deserialize(json);
-      case 'PageBack': return const PageBack();
-      case 'Descendant': return Descendant.deserialize(json, this);
-      case 'Ancestor': return Ancestor.deserialize(json, this);
-    }
-    throw DriverError('Unsupported search specification type $finderType');
-  }
-}
 
 const List<Type> _supportedKeyValueTypes = <Type>[String, int];
