+++ packages/flutter_driver/test/src/real_tests/find_test.dart
@@ -4,6 +4,7 @@
 
 // @dart = 2.8
 
+import 'package:flutter_driver/driver_extension.dart';
 import 'package:flutter_driver/flutter_driver.dart';
 import 'package:flutter_driver/src/common/find.dart';
 import 'package:mockito/mockito.dart';
