+++ packages/flutter_driver/test/src/real_tests/stubs/stub_finder_extension.dart
@@ -3,9 +3,9 @@
 // found in the LICENSE file.
 
 import 'package:flutter/src/widgets/framework.dart';
-import 'package:flutter_driver/driver_extension.dart';
-import 'package:flutter_driver/src/common/create_finder_factory.dart';
 import 'package:flutter_test/src/finders.dart';
+import 'package:flutter_driver/driver_extension.dart';
+import 'package:flutter_driver/src/common/handler_factory.dart';
 import 'package:flutter_driver/src/common/find.dart';
 
 import 'stub_finder.dart';
